pyproject.toml
[tool.poetry]
name = "58-minionclassifier"
version = "0.1.0"
description = ""
authors = ["mojoee <47809126+mojoee@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "58_minionclassifier"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
beautifulsoup4 = "^4.12.2"
requests = "^2.28.2"
tensorflow-macos = "^2.12.0"
pillow = "^9.5.0"
matplotlib = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
